requirements.txt
django
pyyaml
django-compressor
djangorestframework
